requirements.txt
fastapi
uvicorn
langchain
langchain-experimental
langchain-community
streamlit
pandas
matplotlib
seaborn
pillow
requests
pydantic
transformers
torch